// src/api/types/OrganizationUnitDto.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { PositionDto } from './PositionDto'

export type OrganizationUnitDto = {
    referenceId?: string;
    path?: string;
    name?: string;
    code?: string;
    description?: string | null;
    /**
     * @description
     * Format: `date-time`
     * @type string | undefined
    */
    createdOnUtc?: string;
    createdBy?: string | null;
    /**
     * @description
     * Format: `date-time`
     * @type string | undefined
    */
    lastModifiedOnUtc?: string | null;
    lastModifiedBy?: string | null;
    isDeleted?: boolean;
    /**
     * @description
     * Format: `date-time`
     * @type string | undefined
    */
    deletedOnUtc?: string | null;
    deletedBy?: string | null;
    children?: OrganizationUnitDto[];
    descendants?: OrganizationUnitDto[];
    positions?: PositionDto[];
};